+++ CHANGELOG.md
@@ -17,6 +17,7 @@
    * CHANGED: implement valhalla_build_elevation in python and add more --from-geojson & --from-graph options [#3318](https://github.com/valhalla/valhalla/pull/3318)
    * ADDED: Add boolean parameter to clear memory for edge labels from thor. [#2789](https://github.com/valhalla/valhalla/pull/2789)
    * CHANGED: Do not create statsd client in workers if it is not configured [#3394](https://github.com/valhalla/valhalla/pull/3394)
+   * ADDED: Add potential tile to all GetGraphTile calls [#2780](https://github.com/valhalla/valhalla/pull/2780)
 
 ## Release Date: 2021-10-07 Valhalla 3.1.4
 * **Removed**